fix: Automate manual test and fix missing locks
Running the manual test with the race detector reviewed more places in
scope.go that are missing synchronization to be safe for concurrent use.

This commit replaces the manual test file with an automated test and
adds missing calls to Lock() where covered by the test.
